KolamExplorer/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart


python -m uvicorn server:app --reload